refactor de dos archivos
